Allow selection of catalogs
