+++ frontend/vre/catalog/collection.search.view.mustache
@@ -1,0 +1,6 @@
+<div class="page-header" id="title-HPB">
+    <h2>{{name}}</h2>
+</div>
+<div class="well well-lg"  id="HPB-info">
+    {{description}}
+</div>
