+++ frontend/vre/catalog/select-catalog.view.mustache
@@ -3,6 +3,6 @@
 </a>
 <ul class="dropdown-menu">
     {{#catalogs}}
-        <li> <a href="/catalog/{{identifier}}" title="{{description}}">{{name}}</a></li>
+        <li> <a href="/catalog/{{identifier}}/" title="{{description}}">{{name}}</a></li>
     {{/catalogs}}
 </ul>
